// applets/jellyfin/api/model/itemFields.ts
/**
 * Generated by orval v6.11.1 🍺
 * Do not edit manually.
 * Jellyfin API
 * OpenAPI spec version: 10.8.8
 */

/**
 * Used to control the data that gets attached to DtoBaseItems.
 */
export type ItemFields = (typeof ItemFields)[keyof typeof ItemFields];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ItemFields = {
  AirTime: "AirTime",
  CanDelete: "CanDelete",
  CanDownload: "CanDownload",
  ChannelInfo: "ChannelInfo",
  Chapters: "Chapters",
  ChildCount: "ChildCount",
  CumulativeRunTimeTicks: "CumulativeRunTimeTicks",
  CustomRating: "CustomRating",
  DateCreated: "DateCreated",
  DateLastMediaAdded: "DateLastMediaAdded",
  DisplayPreferencesId: "DisplayPreferencesId",
  Etag: "Etag",
  ExternalUrls: "ExternalUrls",
  Genres: "Genres",
  HomePageUrl: "HomePageUrl",
  ItemCounts: "ItemCounts",
  MediaSourceCount: "MediaSourceCount",
  MediaSources: "MediaSources",
  OriginalTitle: "OriginalTitle",
  Overview: "Overview",
  ParentId: "ParentId",
  Path: "Path",
  People: "People",
  PlayAccess: "PlayAccess",
  ProductionLocations: "ProductionLocations",
  ProviderIds: "ProviderIds",
  PrimaryImageAspectRatio: "PrimaryImageAspectRatio",
  RecursiveItemCount: "RecursiveItemCount",
  Settings: "Settings",
  ScreenshotImageTags: "ScreenshotImageTags",
  SeriesPrimaryImage: "SeriesPrimaryImage",
  SeriesStudio: "SeriesStudio",
  SortName: "SortName",
  SpecialEpisodeNumbers: "SpecialEpisodeNumbers",
  Studios: "Studios",
  BasicSyncInfo: "BasicSyncInfo",
  SyncInfo: "SyncInfo",
  Taglines: "Taglines",
  Tags: "Tags",
  RemoteTrailers: "RemoteTrailers",
  MediaStreams: "MediaStreams",
  SeasonUserData: "SeasonUserData",
  ServiceName: "ServiceName",
  ThemeSongIds: "ThemeSongIds",
  ThemeVideoIds: "ThemeVideoIds",
  ExternalEtag: "ExternalEtag",
  PresentationUniqueKey: "PresentationUniqueKey",
  InheritedParentalRatingValue: "InheritedParentalRatingValue",
  ExternalSeriesId: "ExternalSeriesId",
  SeriesPresentationUniqueKey: "SeriesPresentationUniqueKey",
  DateLastRefreshed: "DateLastRefreshed",
  DateLastSaved: "DateLastSaved",
  RefreshState: "RefreshState",
  ChannelImage: "ChannelImage",
  EnableMediaSourceDisplay: "EnableMediaSourceDisplay",
  Width: "Width",
  Height: "Height",
  ExtraIds: "ExtraIds",
  LocalTrailerCount: "LocalTrailerCount",
  IsHD: "IsHD",
  SpecialFeatureCount: "SpecialFeatureCount",
} as const;